Consolidate auth E2E files and add roadmap feature coverage
