+++ frontend/tests/e2e/utils/authenticate-e2e-user-utils.ts
@@ -14,7 +14,7 @@
  */
 
 import { expect, type BrowserContext, type Page } from "@playwright/test";
-import { createSpoofedSessionCookie } from "tests/e2e/loginUtils";
+import { createSpoofedSessionCookie } from "tests/e2e/utils/loginUtils";
 import playwrightEnv from "tests/e2e/playwright-env";
 import { openMobileNav } from "tests/e2e/playwrightUtils";
 import { performStagingLogin } from "tests/e2e/utils/perform-login-utils";
